:recycle: refactor: split health checks and auth into plugins
