+++ src/lib/fastify/logging.ts
@@ -45,14 +45,3 @@ export const envToLogger: Record<"development" | "production" | "test", FastifyS
     },
   },
 };
-
-/**
- * Extend the Fastify session interface to include our custom properties.
- */
-declare module "fastify" {
-  interface Session {
-    codeVerifier?: string;
-    userId?: string;
-    authenticated: boolean;
-  }
-}
